feat(#65): compile simple examples
