+++ src/main/java/org/eolang/opeo/ast/Literal.java
@@ -23,8 +23,10 @@
  */
 package org.eolang.opeo.ast;
 
+import java.util.Collections;
 import java.util.List;
 import org.eolang.jeo.representation.directives.DirectivesData;
+import org.objectweb.asm.Opcodes;
 import org.xembly.Directive;
 
 /**
@@ -53,7 +55,13 @@ public final class Literal implements AstNode {
 
     @Override
     public List<Opcode> opcodes() {
-        throw new UnsupportedOperationException("Not implemented yet");
+        if (this.object instanceof Integer) {
+            return Collections.singletonList(
+                Literal.opcode((Integer) this.object)
+            );
+        } else {
+            throw new UnsupportedOperationException("Not implemented yet");
+        }
     }
 
     @Override
@@ -65,5 +73,38 @@ public final class Literal implements AstNode {
             result = this.object.toString();
         }
         return result;
+    }
+
+    /**
+     * Convert integer into an opcode.
+     * @param value Integer value.
+     * @return Opcode.
+     */
+    private static Opcode opcode(final int value) {
+        final Opcode res;
+        switch (value) {
+            case 0:
+                res = new Opcode(Opcodes.ICONST_0);
+                break;
+            case 1:
+                res = new Opcode(Opcodes.ICONST_1);
+                break;
+            case 2:
+                res = new Opcode(Opcodes.ICONST_2);
+                break;
+            case 3:
+                res = new Opcode(Opcodes.ICONST_3);
+                break;
+            case 4:
+                res = new Opcode(Opcodes.ICONST_4);
+                break;
+            case 5:
+                res = new Opcode(Opcodes.ICONST_5);
+                break;
+            default:
+                res = new Opcode(Opcodes.BIPUSH, value);
+                break;
+        }
+        return res;
     }
 }
