+++ src/main/java/org/eolang/opeo/compilation/OpeoNodes.java
@@ -28,9 +28,13 @@ import java.util.Collections;
 import java.util.List;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
+import org.eolang.jeo.representation.xmir.XmlInstruction;
 import org.eolang.jeo.representation.xmir.XmlNode;
+import org.eolang.opeo.ast.Add;
 import org.eolang.opeo.ast.AstNode;
+import org.eolang.opeo.ast.Literal;
 import org.eolang.opeo.ast.Opcode;
+import org.eolang.opeo.jeo.JeoInstruction;
 import org.objectweb.asm.Opcodes;
 import org.xembly.Xembler;
 
@@ -92,58 +96,30 @@ public final class OpeoNodes {
         //  to bytecode instructions.
         //  We should generate AST first and then compile it to bytecode instructions.
         //  Don't forget to add unit tests.
-        if (node.hasAttribute("base", ".plus")) {
-            final List<XmlNode> inner = node.children().collect(Collectors.toList());
-            result = Stream.of(
-                OpeoNodes.opcode(new HexString(inner.get(0).text()).decodeAsInt()),
-                OpeoNodes.opcode(new HexString(inner.get(1).text()).decodeAsInt()),
-                new Opcode(Opcodes.IADD)
-                )
-                .map(Opcode::toXmir)
-                .map(Xembler::new)
-                .map(Xembler::xmlQuietly)
-                .map(XmlNode::new)
-                .collect(Collectors.toList());
-        } else {
-            result = Collections.singletonList(node);
-        }
-        return result;
+        return OpeoNodes.node(node).opcodes()
+            .stream()
+            .map(Opcode::toXmir)
+            .map(Xembler::new)
+            .map(Xembler::xmlQuietly)
+            .map(XmlNode::new)
+            .collect(Collectors.toList());
     }
 
-
-
-
-
-    /**
-     * Convert integer into an opcode.
-     * @param value Integer value.
-     * @return Opcode.
-     */
-    private static Opcode opcode(final int value) {
-        final Opcode res;
-        switch (value) {
-            case 0:
-                res = new Opcode(Opcodes.ICONST_0);
-                break;
-            case 1:
-                res = new Opcode(Opcodes.ICONST_1);
-                break;
-            case 2:
-                res = new Opcode(Opcodes.ICONST_2);
-                break;
-            case 3:
-                res = new Opcode(Opcodes.ICONST_3);
-                break;
-            case 4:
-                res = new Opcode(Opcodes.ICONST_4);
-                break;
-            case 5:
-                res = new Opcode(Opcodes.ICONST_5);
-                break;
-            default:
-                res = new Opcode(Opcodes.BIPUSH, value);
-                break;
+    private static AstNode node(final XmlNode node) {
+        if (node.hasAttribute("base", ".plus")) {
+            final List<XmlNode> inner = node.children().collect(Collectors.toList());
+            final AstNode left = OpeoNodes.node(inner.get(0));
+            final AstNode right = OpeoNodes.node(inner.get(1));
+            return new Add(left, right);
+        } else if (node.hasAttribute("base", "opcode")) {
+            final XmlInstruction instruction = new XmlInstruction(node.node());
+            return new Opcode(instruction.opcode(), instruction.operands());
+        } else if (node.hasAttribute("base", "int")) {
+            return new Literal(new HexString(node.text()).decodeAsInt());
+        } else {
+            throw new IllegalArgumentException(
+                String.format("Can't recognize node: %n%s%n", node)
+            );
         }
-        return res;
     }
 }
